+++ CHANGELOG.md
@@ -4,7 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- DF-978 Added overwrite option for package import.
+
 ### Changed
+- DF-967 Made the error message 'No record(s) detected in request.' more verbose.
+
 ### Fixed
 
 ## [0.7.0] - 2017-01-16